Move nano::account_info in to its own file.
