+++ nano/secure/common.cpp
@@ -218,65 +218,6 @@ nano::keypair::keypair (std::string const & prv_a)
 	ed25519_publickey (prv.bytes.data (), pub.bytes.data ());
 }
 
-nano::account_info::account_info (nano::block_hash const & head_a, nano::account const & representative_a, nano::block_hash const & open_block_a, nano::amount const & balance_a, nano::seconds_t modified_a, uint64_t block_count_a, nano::epoch epoch_a) :
-	head (head_a),
-	representative (representative_a),
-	open_block (open_block_a),
-	balance (balance_a),
-	modified (modified_a),
-	block_count (block_count_a),
-	epoch_m (epoch_a)
-{
-}
-
-bool nano::account_info::deserialize (nano::stream & stream_a)
-{
-	auto error (false);
-	try
-	{
-		nano::read (stream_a, head.bytes);
-		nano::read (stream_a, representative.bytes);
-		nano::read (stream_a, open_block.bytes);
-		nano::read (stream_a, balance.bytes);
-		nano::read (stream_a, modified);
-		nano::read (stream_a, block_count);
-		nano::read (stream_a, epoch_m);
-	}
-	catch (std::runtime_error const &)
-	{
-		error = true;
-	}
-
-	return error;
-}
-
-bool nano::account_info::operator== (nano::account_info const & other_a) const
-{
-	return head == other_a.head && representative == other_a.representative && open_block == other_a.open_block && balance == other_a.balance && modified == other_a.modified && block_count == other_a.block_count && epoch () == other_a.epoch ();
-}
-
-bool nano::account_info::operator!= (nano::account_info const & other_a) const
-{
-	return !(*this == other_a);
-}
-
-size_t nano::account_info::db_size () const
-{
-	debug_assert (reinterpret_cast<uint8_t const *> (this) == reinterpret_cast<uint8_t const *> (&head));
-	debug_assert (reinterpret_cast<uint8_t const *> (&head) + sizeof (head) == reinterpret_cast<uint8_t const *> (&representative));
-	debug_assert (reinterpret_cast<uint8_t const *> (&representative) + sizeof (representative) == reinterpret_cast<uint8_t const *> (&open_block));
-	debug_assert (reinterpret_cast<uint8_t const *> (&open_block) + sizeof (open_block) == reinterpret_cast<uint8_t const *> (&balance));
-	debug_assert (reinterpret_cast<uint8_t const *> (&balance) + sizeof (balance) == reinterpret_cast<uint8_t const *> (&modified));
-	debug_assert (reinterpret_cast<uint8_t const *> (&modified) + sizeof (modified) == reinterpret_cast<uint8_t const *> (&block_count));
-	debug_assert (reinterpret_cast<uint8_t const *> (&block_count) + sizeof (block_count) == reinterpret_cast<uint8_t const *> (&epoch_m));
-	return sizeof (head) + sizeof (representative) + sizeof (open_block) + sizeof (balance) + sizeof (modified) + sizeof (block_count) + sizeof (epoch_m);
-}
-
-nano::epoch nano::account_info::epoch () const
-{
-	return epoch_m;
-}
-
 nano::pending_info::pending_info (nano::account const & source_a, nano::amount const & amount_a, nano::epoch epoch_a) :
 	source (source_a),
 	amount (amount_a),
