+++ nano/secure/common.hpp
@@ -99,29 +99,6 @@ public:
 };
 
 /**
- * Latest information about an account
- */
-class account_info final
-{
-public:
-	account_info () = default;
-	account_info (nano::block_hash const &, nano::account const &, nano::block_hash const &, nano::amount const &, nano::seconds_t modified, uint64_t, epoch);
-	bool deserialize (nano::stream &);
-	bool operator== (nano::account_info const &) const;
-	bool operator!= (nano::account_info const &) const;
-	size_t db_size () const;
-	nano::epoch epoch () const;
-	nano::block_hash head{ 0 };
-	nano::account representative{};
-	nano::block_hash open_block{ 0 };
-	nano::amount balance{ 0 };
-	/** Seconds since posix epoch */
-	nano::seconds_t modified{ 0 };
-	uint64_t block_count{ 0 };
-	nano::epoch epoch_m{ nano::epoch::epoch_0 };
-};
-
-/**
  * Information on an uncollected send
  */
 class pending_info final
