+++ nano/store/account.hpp
@@ -2,12 +2,14 @@
 
 #include <nano/lib/numbers.hpp>
 #include <nano/store/component.hpp>
+#include <nano/store/db_val_impl.hpp>
 #include <nano/store/iterator.hpp>
 
 #include <functional>
 
 namespace nano
 {
+class account_info;
 class block_hash;
 }
 namespace nano::store
